Be consistent with variable names and use size_t for array indices in fit.c

+++ fit.h
@@ -8,6 +8,6 @@
 #ifndef FIT_H
 #define FIT_H
 
-int fit_sinusoids(double *x, double *y, double *e, int c, double *freqs, int numFreqs, double *amplitudes);
-int fit_polynomial(double *x, double *y, double *e, int c, double *coeffs, int degree);
+int fit_sinusoids(double *x, double *y, double *e, size_t c, double *freqs, size_t numFreqs, double *amplitudes);
+int fit_polynomial(double *x, double *y, double *e, size_t c, double *coeffs, size_t degree);
 #endif
